Fixed: Shader & UV input. Applied still image capture from webcam. Fixed bug with PlaneRect constructor (p01 was swapped with p10).
Must: Remove Intersector class, BoundCalc class, and TouchScreenManager. They contribute nothing in this project.

Unity issues: Unity crashes if I try to update a live feed from a webcam and stick it in a shader. The program is primed to make this transition in InterpolationManager, but doesn't currently implement it.

Unresolved Issues: Moving the viewing camera or the associated clip plane seems to mess up either the mesh of the plane constructed or the image presented. Refer to ViewingCamManager or AggregateClipPlane for more info.